feat: refactor verification process and enhance project structure
- Removed the orchestrator module and replaced it with a workflow coordinator for improved modularity.
- Updated README to reflect changes in core components and added details about new services.
- Enhanced temporal analysis error handling for better debugging.
- Improved fact-checking service to provide progress updates more efficiently.
- Updated frontend components to utilize new hooks for WebSocket and verification state management.
- Cleaned up unused files and optimized file validation logic.

+++ backend/app/schemas.py
@@ -43,7 +43,7 @@ class VerificationResponse(BaseModel):
     identified_claims: Optional[List[str]] = None
     verdict: Optional[str] = None  # true, partially_true, false, ironic
     justification: Optional[str] = None
-    confidence_score: Optional[int] = None
+    confidence_score: Optional[float] = None
     processing_time_seconds: Optional[int] = None
     user_reputation: Optional[UserReputation] = None
     warnings: Optional[List[str]] = None
